requirements_new.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.9.1
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
audioop-lts==0.2.2
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
crispy-bootstrap5==2025.6
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
discord==2.3.2
discord.py==2.5.2
Django==4.2.26
django-cors-headers==4.8.0
django-crispy-forms==2.4
django-ninja==1.5.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
keras==3.12.0
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lark==1.3.1
libclang==18.1.1
llvmlite==0.44.0
marisa-trie==1.3.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.3
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.4
murmurhash==1.0.13
mysql==0.0.3
mysqlclient==2.2.7
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
opt_einsum==3.4.0
optree==0.18.0
outcome==1.3.0.post0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
preshed==3.0.10
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.33.1
psutil==7.1.3
pure_eval==0.2.3
py-cord==2.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.2
PyNaCl==1.5.0
pyparsing==3.2.5
PySocks==1.7.1
pytest==9.0.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-json-logger==4.0.0
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.7.34
reportlab==4.4.5
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.29.0
scipy==1.16.3
selenium==4.38.0
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
terminado==0.18.1
thinc==8.3.6
tiktoken==0.11.0
tinycss2==1.4.0
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
webcolors==25.10.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
wheel==0.45.1
whisper==1.1.10
wrapt==1.17.3
wsproto==1.3.2
yarl==1.20.1
yt-dlp==2025.8.11
